restapi_demo/requirements.txt
apipkg==1.5
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
awscli==1.16.108
boto==2.49.0
boto3==1.9.98
botocore==1.12.98
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.5
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-filter==2.1.0
django-rest-auth==0.9.3
django-storages==1.7.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
docutils==0.14
execnet==1.5.0
idna==2.8
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
more-itertools==5.0.0
oauthlib==3.0.1
Pillow==5.4.1
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.7.0
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
pytest==4.2.0
pytest-django==3.4.7
pytest-forked==1.0.1
pytest-xdist==1.26.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
reverse==0.1.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
urllib3==1.24.1
